+++ rpcs3/emucore.vcxproj.filters
@@ -1063,6 +1063,9 @@
     <ClCompile Include="Crypto\decrypt_binaries.cpp">
       <Filter>Crypto</Filter>
     </ClCompile>
+    <ClCompile Include="Emu\RSX\RSXZCULL.cpp">
+      <Filter>Emu\GPU\RSX</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Crypto\aes.h">
@@ -2115,6 +2118,9 @@
     </ClInclude>
     <ClInclude Include="Crypto\decrypt_binaries.h">
       <Filter>Crypto</Filter>
+    </ClCompile>
+    <ClInclude Include="Emu\RSX\RSXZCULL.h">
+      <Filter>Emu\GPU\RSX</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
